К названию тестовых пакетов добавлен check
