+++ tests/com.e1c.v8codestyle.md.itests/src/com/e1c/v8codestyle/md/check/itests/CommonModuleNameGlobalTest.java
@@ -11,7 +11,7 @@
  *     1C-Soft LLC - initial API and implementation
  *     Aleksandr Kapralov - issue #14
  *******************************************************************************/
-package com.e1c.v8codestyle.md.itests;
+package com.e1c.v8codestyle.md.check.itests;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -32,32 +32,32 @@ import com._1c.g5.v8.dt.metadata.mdclass.CommonModule;
 import com._1c.g5.v8.dt.metadata.mdclass.ReturnValuesReuse;
 import com._1c.g5.v8.dt.validation.marker.Marker;
 import com.e1c.g5.v8.dt.testing.check.CheckTestBase;
-import com.e1c.v8codestyle.md.check.CommonModuleNameClientServer;
+import com.e1c.v8codestyle.md.check.CommonModuleNameGlobal;
 import com.e1c.v8codestyle.md.check.CommonModuleType;
 
 /**
- * Tests for {@link CommonModuleNameClientServer} check.
+ * Tests for {@link CommonModuleNameGlobal} check.
  *
  * @author Dmitriy Marmyshev
  *
  */
-public class CommonModuleNameClientServerTest
+public class CommonModuleNameGlobalTest
     extends CheckTestBase
 {
 
-    private static final String CHECK_ID = "common-module-name-client-server";
+    private static final String CHECK_ID = "common-module-name-global";
 
     private static final String PROJECT_NAME = "CommonModuleName";
 
     private static final String MODULE_DEFAULT_FQN = "CommonModule.CommonModuleName";
 
     @Test
-    public void testCommonModuleNameClientServer() throws Exception
+    public void testCommonModuleNameClientGlobal() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_GLOBAL,
             ReturnValuesReuse.DONT_USE, null);
 
         long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
@@ -66,14 +66,14 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerCorrect() throws Exception
+    public void testCommonModuleNameClientGlobalCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.CommonModuleClientServer";
+        String fqn = "CommonModule.CommonModuleClientGlobal";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_GLOBAL,
             ReturnValuesReuse.DONT_USE, fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
@@ -82,14 +82,14 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerWithPostfixCorrect() throws Exception
+    public void testCommonModuleNameClientGlobalWithPostfixCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.CommonModuleClientServerPredefined";
+        String fqn = "CommonModule.CommonModuleClientGlobalPredefined";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_GLOBAL,
             ReturnValuesReuse.DONT_USE, fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
@@ -98,14 +98,14 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerWithPrefixIncorrect() throws Exception
+    public void testCommonModuleNameClientGlobalWithPrefixIncorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.ClientServerCommonModule";
+        String fqn = "CommonModule.GlobalCommonModule";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_GLOBAL,
             ReturnValuesReuse.DONT_USE, fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
@@ -114,12 +114,88 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerReturnValueReuseCorrect() throws Exception
+    public void testCommonModuleNameClientGlobalReturnValueReuseCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_GLOBAL,
+            ReturnValuesReuse.DONT_USE, null);
+
+        long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
+        Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
+        assertNotNull(marker);
+    }
+
+    @Test
+    public void testCommonModuleNameServerGlobal() throws Exception
+    {
+        IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
+        assertNotNull(dtProject);
+
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_SERVER_GLOBAL,
+            ReturnValuesReuse.DONT_USE, null);
+
+        long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
+        Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
+        assertNotNull(marker);
+    }
+
+    @Test
+    public void testCommonModuleNameServerGlobalCorrect() throws Exception
+    {
+        IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
+        assertNotNull(dtProject);
+
+        String fqn = "CommonModule.CommonModuleServerGlobal";
+
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_SERVER_GLOBAL,
+            ReturnValuesReuse.DONT_USE, fqn);
+
+        long id = getTopObjectIdByFqn(fqn, dtProject);
+        Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
+        assertNull(marker);
+    }
+
+    @Test
+    public void testCommonModuleNameServerGlobalWithPostfixCorrect() throws Exception
+    {
+        IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
+        assertNotNull(dtProject);
+
+        String fqn = "CommonModule.CommonModuleServerGlobalPredefined";
+
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_SERVER_GLOBAL,
+            ReturnValuesReuse.DONT_USE, fqn);
+
+        long id = getTopObjectIdByFqn(fqn, dtProject);
+        Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
+        assertNull(marker);
+    }
+
+    @Test
+    public void testCommonModuleNameServerGlobalWithPrefixIncorrect() throws Exception
+    {
+        IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
+        assertNotNull(dtProject);
+
+        String fqn = "CommonModule.GlobalCommonModule";
+
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_SERVER_GLOBAL,
+            ReturnValuesReuse.DONT_USE, fqn);
+
+        long id = getTopObjectIdByFqn(fqn, dtProject);
+        Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
+        assertNotNull(marker);
+    }
+
+    @Test
+    public void testCommonModuleNameServerGlobalReturnValueReuseCorrect() throws Exception
+    {
+        IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
+        assertNotNull(dtProject);
+
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_SERVER_GLOBAL,
             ReturnValuesReuse.DURING_SESSION, null);
 
         long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
