+++ tests/com.e1c.v8codestyle.ql.itests/src/com/e1c/v8codestyle/ql/check/itests/CastToMaxNumberTest.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     1C-Soft LLC - initial API and implementation
  *******************************************************************************/
-package com.e1c.v8codestyle.ql.itests;
+package com.e1c.v8codestyle.ql.check.itests;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
